Refactored controller.
